[+][Feat] Implement an Abstract Class for Commands classes. It will handle the send to endpoint methods, push to history methods and the ack handling. File created : src/Command/ACommand.hpp

+++ src/Command/Factory/Factory.hh
@@ -50,11 +50,11 @@ namespace Rtype
                 Factory();
                 ~Factory() = default;
 
-                std::unique_ptr<ICommand> createCommand(int cmd_category, int cmd_index);
+                std::unique_ptr<ICommand> createCommand(uint8_t cmd_category, uint8_t cmd_index);
 
             protected:
             private:
-                std::map<std::pair<int, int>, std::function<std::unique_ptr<ICommand>()>> _commandMap;
+                std::map<std::pair<uint8_t, uint8_t>, std::function<std::unique_ptr<ICommand>()>> _commandMap;
         };
     }
 }
